Added notification for canceling a training.
